+++ src/storage/index.ts
@@ -15,6 +15,7 @@ export * from "./utils/contain-path";
 export * from "./utils/mime";
 export * from "./utils/safe-fetch";
 export * from "./utils/storage-error";
+export * from "./utils/storage-not-initialized-error";
 
 // Drivers
 export * from "./drivers/cloud-driver";